Add pagination and suggested changes
