+++ app/views/_pagination.html.erb
@@ -1,0 +1,17 @@
+<nav class="d-flex justify-content-center">
+  <ul class="pagination">
+    <% if @page > 1 %>
+      <li class="page-item">
+        <a class="page-link" href="<%= user_posts_path(@user, page: @page - 1) %>">Previous</a>
+      </li>
+    <% end %>
+    <li class="page-item">
+      <a class="page-link" href="#"><%= @page %></a>
+    </li>
+    <% if @page < @total_pages %>
+      <li class="page-item">
+        <a class="page-link" href="<%= user_posts_path(@user, page: @page + 1) %>">Next</a>
+      </li>
+    <% end %>
+  </ul>
+</nav>
