+++ app/views/posts/index.html.erb
@@ -1,5 +1,5 @@
 <%= render '/user_card', user: @user %>
-<% @user.posts.sort.each do |post| %>
+<% @posts.sort.each do |post| %>
   <section class="mb-5">
     <%= link_to render('/post_card', post:), user_post_path(@user.id, post.id) %>
     <div class="px-3">
@@ -8,4 +8,7 @@
       <% end %>
     </div>
   </section>
+<% end %>
+<% if @pagination %>
+  <%= render '/pagination' %>
 <% end %>
